rest_server/requirements.txt
aniso8601==9.0.1
blinker==1.6.2
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.1.0
cycler==0.11.0
Flask==2.3.2
Flask-RESTful==0.3.10
fonttools==4.40.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.1
jsonpickle==3.0.1
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
more-itertools==9.1.0
music21==9.1.0
networkx==3.1
numpy==1.25.0
packaging==23.1
Pillow==10.0.0
py-dactyl==2.0.2
pydactyl==0.5
PyMySQL==1.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
TatSu==5.8.3
urllib3==2.0.3
webcolors==1.13
Werkzeug==2.3.6
